backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.9
requests>=2.31.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=42.0.0
openpyxl>=3.1.0
python-dateutil>=2.8.0
PyJWT>=2.8.0
bcrypt>=4.1.0
